requirements.txt
pypdf>=6.0.0
pypdfium2>=5.6.0
pillow>=10.0.0
pytesseract>=0.3.13
pandas>=2.0.0
openpyxl>=3.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
